fix: move freeze out of _commands
